fix(db): replace numeric(14,8) with numeric(16,8) for monetary columns
numeric(14,8) only allows 6 integer digits (max ~999,999) but existing
balance_usd data reaches ~24M. Widening to (16,8) gives 8 integer digits
(max ~99.9M). Deleted the broken 0018/0019 migrations and regenerated.

+++ packages/db/src/schema/client.ts
@@ -47,19 +47,19 @@ export const clients = pgTable("clients", {
 	email: varchar("email", { length: 255 }).notNull().unique(),
 	passwordHash: varchar("password_hash", { length: 255 }).notNull(),
 	balanceUsd: numeric("balance_usd", {
-		precision: 14,
+		precision: 16,
 		scale: 8,
 		mode: "number",
 	})
 		.notNull()
 		.default(0),
 	monthlyUsageLimit: numeric("monthly_usage_limit", {
-		precision: 14,
+		precision: 16,
 		scale: 8,
 		mode: "number",
 	}),
 	usageAlertThresholdUsd: numeric("usage_alert_threshold_usd", {
-		precision: 14,
+		precision: 16,
 		scale: 8,
 		mode: "number",
 	}),
