+++ packages/db/src/schema/usage.ts
@@ -25,7 +25,7 @@ export const usageRecords = pgTable(
 		type: usageTypeEnum("type").notNull().default("message"),
 		tokensUsed: integer("tokens_used").notNull(),
 		costUsd: numeric("cost_usd", {
-			precision: 14,
+			precision: 16,
 			scale: 8,
 			mode: "number",
 		}).notNull(),
